feat(owner): add OV re-registration window checks
This adds the ability to trigger a rendezvous re-registration
when an OV is close to its de-registration timeout. A re-registration
window is added so that when the remaining registration period
of the OV reaches that window, a re-registration is triggered.
This feature is only available when using database storage
drivers.

Signed-off-by: Irene Diez <idiez@redhat.com>